Fix improper function signature

+++ jre_lwjgl3glfw/src/main/java/org/lwjgl/glfw/CallbackBridge.java
@@ -53,7 +53,7 @@ public class CallbackBridge {
     }
     
 	// Called from Android side
-	public static void receiveCallback(int type, float i1, float i2, int i3, int i4) {
+	public static void receiveCallback(int type, int i1, int i2, int i3, int i4) {
        /*
         if (INPUT_DEBUG_ENABLED) {
             System.out.println("LWJGL GLFW Callback received type=" + Integer.toString(type) + ", data=" + i1 + ", " + i2 + ", " + i3 + ", " + i4);
